1.using c library instead of cpp of netcdf to avoid conflicts from ncdfFlow package 2.change include and lib flags in Makevars file to make sure ncdfFlow and flowWorkspace use the same version of cdf library for cdf IO operation 3.implement plotGate and getData R APIs

+++ src/R_GatingHierarchy.cpp
@@ -140,7 +140,7 @@ BEGIN_RCPP
 		case POLYGONGATE:
 			{
 				vector<coordinate> vert=g->getVertices();
-				vector<float> x,y;
+				vector<double> x,y;
 				for(vector<coordinate>::iterator it=vert.begin();it!=vert.end();it++)
 				{
 					x.push_back(it->x);
